src/main/java/tr/org/linux/opencourseorganizer/client/DesktopClientFactory.java
package tr.org.linux.opencourseorganizer.client;

import com.google.gwt.place.shared.PlaceController;
import com.google.web.bindery.event.shared.EventBus;
import com.google.web.bindery.event.shared.SimpleEventBus;

public class DesktopClientFactory implements ClientFactory {

	private final App app = new App();
	private final SimpleEventBus eventBus = new SimpleEventBus();
	private final PlaceController placeController = new PlaceController(eventBus);

	@Override
	public App getApp() {
		return app;
	}

	@Override
	public EventBus getEventBus() {
		return eventBus;
	}

	@Override
	public PlaceController getPlaceController() {
		return placeController;
	}

}
